+++ statistics-service/src/main/java/com/timeline/stats/dto/PlacesVisitedDTO.java
@@ -1,11 +1,13 @@
 package com.timeline.stats.dto;
 
 import java.util.Arrays;
+import java.util.Map;
 import java.util.Set;
 
 public class PlacesVisitedDTO {
   public int count;
   public String[] zipcodes;
+  public Map<String, PlaceInfoDTO> data;
 
   public PlacesVisitedDTO() {
   }
@@ -14,5 +16,10 @@ public class PlacesVisitedDTO {
     this.count = zipcodes.size();
     this.zipcodes = zipcodes.toArray(new String[0]);
     Arrays.sort(this.zipcodes);
+  }
+
+  public PlacesVisitedDTO(Set<String> zipcodes, Map<String, PlaceInfoDTO> data) {
+    this(zipcodes);
+    this.data = data;
   }
 }